Allow count to be used in sscan, hscan and zscan
The sscan, hscan and zscan methods all take 4 arguments:

iterator
field
pattern
count

They simply wrap the __call function with the appropriate redis method
name. Count is always passed as 4th element in the argument array, eg:
$args[3].

The code in __call will check for a non-empty value in $args[3], and
then use $args[4] as the value for COUNT. As $args[4] can never be set,
this is always null, and causes an array reference warning, as well as
a syntax error with the sscan/hscan/zscan method.

This update takes the count value from $args[3], where it is set by the
wrapper methods.

Additionally, the tests have been updated to include a count so that
this kind of issue will be picked up in the future if there are any
regressions.